+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Made strafing slightly slower
 - Food now has limited regeneration strength but longer duration.
 - Harvester boss now has new abilities and AI
+- Death particles and SFX
 
 ### Removed